random hash implementation
