+++ packages/lambda/src/api/presign-url.ts
@@ -1,8 +1,8 @@
 import {GetObjectCommand, HeadObjectCommand} from '@aws-sdk/client-s3';
 import {getSignedUrl} from '@aws-sdk/s3-request-presigner';
+import {validateBucketName} from '@remotion/serverless/client';
 import type {AwsRegion} from '../regions';
 import {getS3Client} from '../shared/get-s3-client';
-import {validateBucketName} from '../shared/validate-bucketname';
 import {validatePresignExpiration} from '../shared/validate-presign-expiration';
 
 export type PresignUrlInput<CheckIfObjectExists extends boolean = boolean> = {
