+++ packages/lambda/src/functions/aws-implementation.ts
@@ -6,6 +6,7 @@ import {lambdaDeleteFileImplementation} from '../io/delete-file';
 import {lambdaLsImplementation} from '../io/list-objects';
 import type {AwsRegion} from '../regions';
 import {applyLifeCyleOperation} from '../shared/lifecycle-rules';
+import {randomHashImplementation} from '../shared/random-hash';
 import {getCurrentRegionInFunctionImplementation} from './helpers/get-current-region';
 
 if (
@@ -32,4 +33,5 @@ export const awsImplementation: ProviderSpecifics<AwsRegion> = {
 	listObjects: lambdaLsImplementation,
 	deleteFile: lambdaDeleteFileImplementation,
 	bucketExists: bucketExistsInRegionImplementation,
+	randomHash: randomHashImplementation,
 };
