+++ packages/serverless/src/client.ts
@@ -1,5 +1,6 @@
 export {Await} from './await';
 export {
+	REMOTION_BUCKET_PREFIX,
 	ServerlessCodec,
 	ServerlessPayload,
 	ServerlessPayloads,
@@ -17,6 +18,13 @@ export {
 	type SerializedInputProps,
 	type WebhookOption,
 } from './constants';
+export {
+	GetOrCreateBucketInput,
+	GetOrCreateBucketOutput,
+	internalGetOrCreateBucket,
+} from './get-or-create-bucket';
 
 export {inputPropsKey, resolvedPropsKey} from './input-props-keys';
+export {makeBucketName} from './make-bucket-name';
 export {streamToString} from './stream-to-string';
+export {validateBucketName} from './validate-bucket-name';
