+++ packages/serverless/src/make-bucket-name.ts
@@ -1,0 +1,8 @@
+import {REMOTION_BUCKET_PREFIX} from './constants';
+import {randomHash} from './random-hash';
+
+export const makeBucketName = <Region extends string>(region: Region) => {
+	return `${REMOTION_BUCKET_PREFIX}${region.replace(/-/g, '')}-${randomHash({
+		randomInTests: false,
+	})}`;
+};
