+++ packages/serverless/src/random-hash.ts
@@ -1,6 +1,7 @@
 const alphabet = 'abcdefghijklmnopqrstuvwxyz0123456789';
 
-export const randomHashImplementation = (): string => {
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const randomHash = (_options?: {randomInTests: boolean}): string => {
 	return new Array(10)
 		.fill(1)
 		.map(() => {
